requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
greenlet==3.1.1
inflection==0.5.1
iniconfig==2.0.0
loguru==0.7.2
packaging==24.1
pluggy==1.5.0
pydantic==1.10.18
PyMySQL==1.1.1
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
uWSGI==2.0.27
